Change type of loop var to size_t to make compiler happy.

+++ src/lib/KeyCodes.cpp
@@ -137,7 +137,7 @@ KeyCode Sexy::GetKeyCodeFromName(const std::string& theKeyName)
 
         std::transform(copy.begin(), copy.end(), copy.begin(), toupper);
 
-    for (int i = 0; i < sizeof(aKeyCodeArray)/sizeof(aKeyCodeArray[0]); i++)    
+    for (size_t i = 0; i < sizeof(aKeyCodeArray)/sizeof(aKeyCodeArray[0]); i++)
           if (strcmp(copy.c_str(), aKeyCodeArray[i].mKeyName) == 0)
             return aKeyCodeArray[i].mKeyCode;   
 
@@ -152,7 +152,7 @@ const std::string Sexy::GetKeyNameFromCode(const KeyCode& theKeyCode)
         return aStr;
     }
 
-    for (int i = 0; i < sizeof(aKeyCodeArray)/sizeof(aKeyCodeArray[0]); i++)    
+    for (size_t i = 0; i < sizeof(aKeyCodeArray)/sizeof(aKeyCodeArray[0]); i++)
         if (theKeyCode == aKeyCodeArray[i].mKeyCode)
             return aKeyCodeArray[i].mKeyName;
